MAR 12 UPDATE
MAR 12 UPDATE:
Basil can now somewhat Queue and Skip due to the delay added. Filtering System updated. As Basil is being tested on two servers. Music Playlist will need to be tested. Basil will most likely be unable to support two servers playing music due to the way it is coded. Will work on it in a future update.